+++ packages/plugin-video/src/toolbar/inline-video-buttons.jsx
@@ -1,8 +1,9 @@
-import { BUTTONS, getModalStyles } from 'wix-rich-content-common';
-import { Modals } from '../modals';
+import { BUTTONS, getModalStyles, decorateComponentWithProps, WixUtils } from 'wix-rich-content-common';
 import { MediaReplaceIcon } from '../icons';
+import VideoSelectionInputModal from './videoSelectionInputModal';
+import { SelectionModalCustomStyle, ExtendedSelectionModalCustomStyle } from './selectionModalCustomStyles';
 
-export default({ t }) => {
+export default ({ t, settings }) => {
   return [
     { keyName: 'sizeSmallCenter', type: BUTTONS.SIZE_SMALL_CENTER, mobile: false },
     { keyName: 'sizeContent', type: BUTTONS.SIZE_CONTENT, mobile: false },
@@ -15,8 +16,15 @@ export default({ t }) => {
       keyName: 'replace',
       type: BUTTONS.EXTERNAL_MODAL,
       icon: MediaReplaceIcon,
-      modalName: Modals.VIDEO_URL_INPUT,
-      modalStyles: getModalStyles({ fullScreen: false }),
+      modalElement: decorateComponentWithProps(VideoSelectionInputModal, settings),
+      modalStyles: getModalStyles({
+        //apply the extended input modal styles if handleFileSelection is avilable in plugin config
+        //& on mobile if enableCustomUploadOnMobile is set to true, otherwise the normal modal styles is applied  
+        customStyles: (!WixUtils.isMobile() || settings.enableCustomUploadOnMobile) && settings.handleFileSelection ?
+          ExtendedSelectionModalCustomStyle :
+          SelectionModalCustomStyle,
+        fullScreen: false
+      }),
       mobile: true,
       tooltipTextKey: 'ReplaceVideoButton_Tooltip',
       t,