+++ packages/plugin-video/src/toolbar/selectionModalCustomStyles.js
@@ -1,0 +1,15 @@
+export const SelectionModalCustomStyle = {
+  content:
+  {
+    maxWidth: '580px',
+    minHeight: '260px'
+  }
+};
+
+export const ExtendedSelectionModalCustomStyle = {
+  content:
+  {
+    maxWidth: '580px',
+    minHeight: '378px'
+  }
+};